+++ src/lib/hooks/index.ts
@@ -7,3 +7,4 @@ export * from "./useSingleMessageProps";
 export * from "./useToast";
 export * from "./useUserKey";
 export * from "./useOpenTxTab";
+export * from "./useIsCurrentPage";